Updated script 01
Added a description of the script.
added commands for the python module to be imported.
Works with a white line to separated different inputs and make it more readable.
Corrected error where mapping.fasta was not found if not existing before.
Corrected error where bowtie fasta did not find the mapping.fasta
Corrected typo on line 68
Allows to loop all sequences and the dispalying the error which ones could not be mapped instead of on occurance.